pyproject.toml
[project]
name = "music-recommendation-engine"
version = "1.0.0"
description = "A FastAPI-based music recommendation system using embeddings and fusion reranking"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "python-multipart>=0.0.6",
    "librosa>=0.10.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "openl3>=0.4.2",
    "pydub>=0.25.1",
    "ffmpeg-python>=0.2.0",
    "mistralai>=1.9.10",
    "openai>=1.106.1",
    "anthropic>=0.66.0",
]

[project.optional-dependencies]
audio = [
    "openl3>=0.4.1",
    "tensorflow>=2.13.0",
    "laion-clap>=1.1.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
include_trailing_comma = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
